Updates to game activity
Added game_player_row and a scroll view to the game activity to add
other players and their current scores. A method updatePlayers can be
called to reload the players. Not sure if it will update the scrollView,
deleting old values or if it will add them to the end.